devbox state monitor
