+++ devbox/common/src/Util.scala
@@ -15,7 +15,6 @@ object Util {
     )
 
   def autoclose[T <: AutoCloseable, V](x: T)(f: T => V) = {
-//    f(x)
     try f(x)
     finally x.close()
   }
